🎨 Group all models into single structure
To simplify db model usage, rather than pass the db connection to every
model individually, we can create a single object that collects all
models into a single object that is shared. This will avoid having to
import every model that needs to be used independently, relying instead
on the context.