HPC-8747: Prevent empty string version tags
`versionTagsString` is a key we use to group
records with same version tags, in order to
update them in bulk.

It is constructed with usage of `Array.prototype.join()`
method, which produces an empty string when
called on empty string, i.e.
`[].join(',')` -> `''`
That means if our record has no previous version
tags, empty string will be used as a key (which
is valid) to group it together with other records
that don't have version tags.

Later, when we do the update, we are splitting
this string key to get back version tags,
but when `split()` is called on an empty string,
that produces `['']`, i.e.
`''.split(',')` -> `['']`.
This array of one element (which is empty string)
is considered as version tags array, to which
we append new tag.

This is how we ended up with empty version tags.