Update Postgres images to 14.12
This is done to match the production version
which is used on AWS RDS, according to the most
recent maintenance by OPS in ticket OPS-10475.

This API depends on v3 API to start the database,
thus integrated compose file is used, instead of
standalone `compose.yaml` file in root of the repo.
Even though this file isn't usually used,
update Postgres version to keep them in sync.