Don't specify `allow` in `dependabot.yml`
Don't specify `allow` in `dependabot.yml`,
because it also applies to security updates.
The `allow` configuration enables us to
"Customize which updates are allowed",
meaning if dependency isn't on the list,
it isn't allowed. So, when security update
attempts to update sub-dependency, which
definitely isn't on the `allow` list,
such update fails.
So, we take the inverse approach, we instead
specify which dependencies NOT to update.